[mod]: check 함수 ViewModel로 분리
